requirements_api.txt
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
